Extract an SpiExtension interface and (de)serialise it into a JsonSpiExtension instance.
