+++ components/proxy/src/main/java/com/hotels/styx/spi/config/SpiExtension.java
@@ -15,83 +15,21 @@
  */
 package com.hotels.styx.spi.config;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
-import com.google.common.base.Objects;
-
-import java.io.IOException;
-
-import static com.fasterxml.jackson.databind.DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES;
-import static com.google.common.base.Objects.toStringHelper;
-import static com.google.common.base.Throwables.propagate;
-import static java.util.Objects.requireNonNull;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 
 /**
  * Factory/configuration block.
  */
-public class SpiExtension {
-    private static final ObjectMapper MAPPER = new ObjectMapper(new YAMLFactory()).configure(FAIL_ON_UNKNOWN_PROPERTIES, false);
+@JsonDeserialize(as = JsonSpiExtension.class)
+public interface SpiExtension {
 
-    private final SpiExtensionFactory factory;
-    private final JsonNode config;
-    private final boolean enabled;
+    SpiExtensionFactory factory();
 
-    @JsonCreator
-    public SpiExtension(@JsonProperty("factory") SpiExtensionFactory factory,
-                        @JsonProperty("config") JsonNode config,
-                        @JsonProperty("enabled") Boolean enabled) {
-        this.factory = requireNonNull(factory, "Factory attribute missing");
-        this.config = requireNonNull(config, "Config attribute missing");
-        this.enabled = enabled == null;
-    }
+    JsonNode config();
 
-    public SpiExtensionFactory factory() {
-        return factory;
-    }
+    boolean enabled();
 
-    public JsonNode config() {
-        return config;
-    }
+    <T> T config(Class<T> configClass);
 
-    public boolean enabled() {
-        return enabled;
-    }
-
-    public <T> T config(Class<T> configClass) {
-        JsonParser parser = config.traverse();
-
-        try {
-            return MAPPER.readValue(parser, configClass);
-        } catch (IOException e) {
-            throw propagate(e);
-        }
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hashCode(factory);
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj) {
-            return true;
-        }
-        if (obj == null || getClass() != obj.getClass()) {
-            return false;
-        }
-        SpiExtension other = (SpiExtension) obj;
-        return Objects.equal(this.factory, other.factory);
-    }
-
-    @Override
-    public String toString() {
-        return toStringHelper(this)
-                .add("factory", factory)
-                .toString();
-    }
 }
