+++ components/proxy/src/test/java/com/hotels/styx/proxy/plugin/PluginsMetadataTest.java
@@ -20,6 +20,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
 import com.google.common.collect.ImmutableMap;
 import com.hotels.styx.common.Pair;
+import com.hotels.styx.spi.config.JsonSpiExtension;
 import com.hotels.styx.spi.config.SpiExtensionFactory;
 import com.hotels.styx.spi.config.SpiExtension;
 import org.testng.annotations.Test;
@@ -67,7 +68,7 @@ public class PluginsMetadataTest {
     }
 
     private SpiExtension pluginMetadata() {
-        return new SpiExtension(factory(), config(), true);
+        return new JsonSpiExtension(factory(), config(), true);
     }
 
     private SpiExtensionFactory factory() {
